fastApi/gps-test/requirements.txt
fastapi==0.68.0
uvicorn==0.15.0
python-multipart==0.0.5
pymongo[srv]==4.0.1
pydantic==1.8.2
python-dotenv==0.19.0
pyyaml==5.4.1
boto3==1.26.137
apscheduler>=3.10.0
